truelayer: Initial integration
Client, auth, and models.

Only the fields that we need at the moment are included in the response
deserialisation and model so that we don't clutter the output.

Automatic detection of sandbox mode based on client ID prefix. Login
details are documented here:

- https://docs.truelayer.com/docs/mock-users

There isn't a supported client for Rust.

Need to run our own short-lived HTTP server for the callback.