sync: Delete existing rules
There's no way to get+update a conditional formatting rule in a single
operation. Add will create duplicates, update requires a rule to already
be there, and there's no way to uniquely identify a rule (e.g. by ID or
name). So we have to delete all of the rules so that our rule is always
up-to-date. At least it happens in a single HTTP request though.

Required some refactoring to pass the `Sheet` up to get the existing
rules from the properties without making an additional request to fetch.